+++ src/geo_activity_playground/features/strava/api_importer.py
@@ -11,7 +11,6 @@ from stravalib import Client
 from stravalib.exc import Fault, ObjectNotFound, RateLimitExceeded
 from tqdm import tqdm
 
-from ...core.activities import ActivityRepository
 from ...core.config import ConfigAccessor
 from ...core.datamodel import DB, Activity, get_or_make_equipment, get_or_make_kind
 from ...core.enrichment import update_and_commit
@@ -134,15 +133,12 @@ def _refresh_activity_names_from_strava_once(config: StravaConfig) -> int:
 
 def import_from_strava_api(
     config_accessor: ConfigAccessor,
-    repository: ActivityRepository,
     strava_begin: str | None = None,
     strava_end: str | None = None,
     source: str | None = None,
 ) -> None:
     try:
-        while try_import_strava(
-            config_accessor, repository, strava_begin, strava_end, source
-        ):
+        while try_import_strava(config_accessor, strava_begin, strava_end, source):
             now = datetime.datetime.now()
             next_quarter = round_to_next_quarter_hour(now)
             seconds_to_wait = (next_quarter - now).total_seconds() + 10
@@ -161,7 +157,6 @@ def import_from_strava_api(
 
 def try_import_strava(
     config_accessor: ConfigAccessor,
-    repository: ActivityRepository,
     strava_begin: str | None = None,
     strava_end: str | None = None,
     source: str | None = None,